+++ src/hardware/apu/ch1.rs
@@ -10,7 +10,7 @@ pub struct Ch1 {
   pub envelope: Envelope,
   freq_bits: u16,
   use_counter: bool,
-  counter: uint,
+  counter: usize,
   pub status: bool
 }
 
@@ -37,7 +37,7 @@ impl Ch1 {
   }
   pub fn write_reg1(&mut self, value: u8) {
     self.wave_duty = FromPrimitive::from_u8((value >> 6) & 0x03).unwrap();
-    self.counter = 64 - (value & 0x3f) as uint;
+    self.counter = 64 - (value & 0x3f) as usize;
   }
   pub fn write_reg3(&mut self, value: u8) {
     self.freq_bits = (self.freq_bits & 0x700) | value as u16;
