+++ src/util/int.rs
@@ -1,7 +1,7 @@
 pub trait IntOps {
   fn isolate_rightmost_one(self) -> Self;
-  fn bit(self, uint) -> Self;
-  fn bit_bool(self, uint) -> bool;
+  fn bit(self, usize) -> Self;
+  fn bit_bool(self, usize) -> bool;
 }
 
 impl IntOps for u8 {
@@ -10,10 +10,10 @@ impl IntOps for u8 {
     // Hacker's Delight 2nd ed, 2-1 Manipulating Rightmost Bits
     self & (-self)
   }
-  fn bit(self, bit: uint) -> u8 {
+  fn bit(self, bit: usize) -> u8 {
     (self >> bit) & 0x01
   }
-  fn bit_bool(self, bit: uint) -> bool {
+  fn bit_bool(self, bit: usize) -> bool {
     self.bit(bit) != 0x00
   }
 }
